requirements.txt
tensorflow>=2.12.0
numpy
pandas
matplotlib
opencv-python
streamlit
Pillow
scikit-learn